more convenient form
Now the series field has the prefix TEN_, CALIBRATE_ or CHECK_ fixed. That way, the user doesn't have to make sure it's correct.
There are two new buttons "M1" and "M2" that ignore the contents of the Capture field and use "M1" and "M2" instead.

There's a dummy class to fake the existence of lights. It's not used in production.