speed and camera properties
Zoom options have been removed from the interface.

An option to open the camera's properties has been added. When the white balance isn't close to 3000, the button turns red and a message is displayed.

M1 and M2 buttons are now ordered.

A lighter shade of blue (more readable) is used in the mode buttons.

Now the camera by default starts with fixed exposure of -5, 0 gain and, if the white balance isn't close to 3000, the camera's properties' window is opened at start.

The image is rotated, flipped, and converted from BGR to RGB with openCV instead of NumPy. Turns out to be faster if later is converted to a PIL image.

If the user sets a prefix in the Serie field, it is ignored.

The message on capture is updated and displayed earlier, the image is saved as a bmp because it is 5x faster than png.