apparent speed up
GUI text disappears with a timer.

When capturing, text appears bold for .1 s.

Images are taken from the camera in another thread.

Images are saved in another thread.

Fixed file name bug when capturing with Enter.

In capture mode, field Captura doesn't increase name. When capturing with MX button it changes to MX.

Captures are again saved as png.

If the calibration doesn't find the pattern it doesn't crashes. It's just slow.