Spinbox & code refactoring
The lights' intensity field is now a Spinbox: the value can be changed with up and down arrows.

Starting and closing logs only display the level and message, not file, function and lineno.

When lights' device is not found, the user is asked with a prettier message.

Some unused utilities have been removed.

When getting the image from the camera, the same array is used always; hopefully this reduces memory use.

Calibration files are now json-valid. The content is generated as the string representation of the Calibration object.