fix prefix save & remove unnecessary button
Calibration and check files are saved with their corresponding prefix.

The "Capturar" button is not shown in "Nuevo usuario" mode, "M1" and "M2" buttons are not shown in calibration and check modes.

If the user code is empty or all zeros, the save message is shown in red.

Now the wait before trying to turn the light is 1.8 s instead of 1 s.